+++ common/src/main/java/net/mehvahdjukaar/polytone/particle/CustomParticlesManager.java
@@ -19,6 +19,7 @@ import net.minecraft.core.RegistryAccess;
 import net.minecraft.core.particles.ParticleType;
 import net.minecraft.core.particles.SimpleParticleType;
 import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.resources.RegistryOps;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.packs.resources.ResourceManager;
 
@@ -39,21 +40,32 @@ public class CustomParticlesManager extends JsonPartialReloader {
         super("custom_particles");
     }
 
+
+    //just gathers the custom models if any are there
     @Override
-    protected void reset() {
+    public void earlyProcess(ResourceManager resourceManager) {
+        var jsons = this.getJsonsInDirectories(resourceManager);
+        for (var j : jsons.entrySet()) {
+            var json = j.getValue();
+            var id = j.getKey();
+            var model = CustomParticleType.CUSTOM_MODEL_ONLY_CODEC.decode(JsonOps.INSTANCE, json)
+                    .getOrThrow(errorMsg -> new IllegalStateException("Could not decode Custom Particle with json id " + id + "\n error: " + errorMsg))
+                    .getFirst();
+            model.ifPresent(Polytone::addCustomModel);
+        }
+    }
+
+    @Override
+    protected void resetWithLevel(boolean isLogOff) {
+        for (var id : customParticleFactories.orderedKeys()) {
+            PlatStuff.unregisterParticleProvider(id);
+            PlatStuff.unregisterDynamic(BuiltInRegistries.PARTICLE_TYPE, id);
+        }
         customParticleFactories.clear();
         for (var v : overwrittenVanillaProviders.entrySet()) {
             PlatStuff.setParticleProvider(v.getKey(), v.getValue());
         }
         overwrittenVanillaProviders.clear();
-    }
-
-    @Override
-    protected void resetWithLevel(boolean logOff) {
-        for (var id : customParticleFactories.orderedKeys()) {
-            PlatStuff.unregisterParticleProvider(id);
-            PlatStuff.unregisterDynamic(BuiltInRegistries.PARTICLE_TYPE, id);
-        }
     }
 
     // not ideal
@@ -71,25 +83,17 @@ public class CustomParticlesManager extends JsonPartialReloader {
 
     @Override
     protected void applyWithLevel(RegistryAccess access, boolean isLogIn) {
-        if(access != null)return;
-        ParticleEngine particleEngine = Minecraft.getInstance().particleEngine;
 
-        for (var c : customParticleFactories.getEntries()) {
-            var factory = c.getValue();
-            var id  = c.getKey();
-            SimpleParticleType type = PlatStuff.makeParticleType(factory.forceSpawns());
-            PlatStuff.registerDynamic(BuiltInRegistries.PARTICLE_TYPE, id, type);
-            particleEngine.register(type, factory);
-        }
     }
 
     @Override
-    protected void process(Map<ResourceLocation, JsonElement> obj, DynamicOps<JsonElement> ops) {
+    protected void parseWithLevel(Map<ResourceLocation, JsonElement> jsons, RegistryOps<JsonElement> ops,
+                                  RegistryAccess access) {
         ParticleEngine particleEngine = Minecraft.getInstance().particleEngine;
 
         Set<CustomParticleType> customTypes = new HashSet<>();
 
-        for (var j : obj.entrySet()) {
+        for (var j : jsons.entrySet()) {
             try {
                 var json = j.getValue();
                 var id = j.getKey();
@@ -125,8 +129,6 @@ public class CustomParticlesManager extends JsonPartialReloader {
             }
         }
 
-        applyWithLevel(null, false);
-
         //initialize recursive stuff
         for (var c : customTypes) {
             for (var d : c.lazyParticles) {
@@ -137,6 +139,15 @@ public class CustomParticlesManager extends JsonPartialReloader {
                 }
             }
             c.lazyParticles = null;
+        }
+
+        // register custom particle types. needs to be here
+        for (var c : customParticleFactories.getEntries()) {
+            var factory = c.getValue();
+            var id  = c.getKey();
+            SimpleParticleType type = PlatStuff.makeParticleType(factory.forceSpawns());
+            PlatStuff.registerDynamic(BuiltInRegistries.PARTICLE_TYPE, id, type);
+            particleEngine.register(type, factory);
         }
     }
 
@@ -149,11 +160,5 @@ public class CustomParticlesManager extends JsonPartialReloader {
 
     public Codec<CustomParticleFactory> byNameCodec() {
         return customParticleFactories;
-    }
-
-    public Iterable<ResourceLocation> getCustomModels() {
-        return customParticleFactories.getValues().stream()
-                .map(CustomParticleFactory::getCustomModel)
-                .filter(Objects::nonNull).collect(Collectors.toSet());
     }
 }