+++ common/src/main/java/net/mehvahdjukaar/polytone/slotify/GuiModifierManager.java
@@ -1,8 +1,6 @@
 package net.mehvahdjukaar.polytone.slotify;
 
 import com.google.gson.JsonElement;
-import com.mojang.serialization.DynamicOps;
-import com.mojang.serialization.JsonOps;
 import it.unimi.dsi.fastutil.ints.Int2ObjectArrayMap;
 import net.mehvahdjukaar.polytone.Polytone;
 import net.mehvahdjukaar.polytone.utils.JsonPartialReloader;
@@ -11,17 +9,17 @@ import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.client.gui.screens.inventory.CreativeModeInventoryScreen;
 import net.minecraft.client.gui.screens.inventory.InventoryScreen;
-import net.minecraft.client.gui.screens.inventory.LecternScreen;
+import net.minecraft.core.RegistryAccess;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.network.chat.contents.TranslatableContents;
+import net.minecraft.resources.RegistryOps;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.inventory.InventoryMenu;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.inventory.Slot;
-import net.minecraft.world.level.material.Fluids;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.*;
@@ -46,7 +44,7 @@ public class GuiModifierManager extends JsonPartialReloader {
     }
 
     @Override
-    protected void reset() {
+    protected void resetWithLevel(boolean logOff) {
         slotsByMenuId.clear();
         slotsByClass.clear();
         slotsByTitle.clear();
@@ -56,10 +54,10 @@ public class GuiModifierManager extends JsonPartialReloader {
     }
 
     @Override
-    protected void process(Map<ResourceLocation, JsonElement> object, DynamicOps<JsonElement> ops) {
+    protected void parseWithLevel(Map<ResourceLocation, JsonElement> jsons, RegistryOps<JsonElement> ops, RegistryAccess access) {
         List<GuiModifier> allModifiers = new ArrayList<>();
 
-        for (var entry : object.entrySet()) {
+        for (var entry : jsons.entrySet()) {
             var json = entry.getValue();
             var id = entry.getKey();
             GuiModifier modifier = GuiModifier.CODEC.decode(ops, json)
@@ -119,7 +117,12 @@ public class GuiModifierManager extends JsonPartialReloader {
             }
 
         }
-        if(!slotsByMenuId.isEmpty() || !slotsByClass.isEmpty() || !slotsByTitle.isEmpty()) {
+
+    }
+
+    @Override
+    protected void applyWithLevel(RegistryAccess access, boolean isLogIn) {
+        if (!slotsByMenuId.isEmpty() || !slotsByClass.isEmpty() || !slotsByTitle.isEmpty()) {
             Polytone.LOGGER.info("Loaded GUI modifiers for: {} {} {} {}", slotsByMenuId.keySet(), slotsByClass.keySet(), byMenuId.keySet(), byClass.keySet());
         }
     }
@@ -231,7 +234,7 @@ public class GuiModifierManager extends JsonPartialReloader {
     }
 
     public boolean maybeChangeColor(AbstractContainerScreen<?> screen, Slot slot, GuiGraphics graphics,
-                                           int x, int y, int offset) {
+                                    int x, int y, int offset) {
         var mod = getSlotModifier(screen, slot);
         if (mod != null && mod.hasCustomColor()) {
             mod.renderCustomHighlight(graphics, x, y, offset);