+++ common/src/main/java/net/mehvahdjukaar/polytone/utils/JsonPartialReloader.java
@@ -1,5 +1,6 @@
 package net.mehvahdjukaar.polytone.utils;
 
+import com.google.common.collect.ImmutableMap;
 import com.google.gson.JsonElement;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.packs.resources.ResourceManager;
@@ -16,6 +17,6 @@ public abstract class JsonPartialReloader extends PartialReloader<Map<ResourceLo
     protected Map<ResourceLocation, JsonElement> prepare(ResourceManager resourceManager) {
         var jsons = this.getJsonsInDirectories(resourceManager);
         this.checkConditions(jsons);
-        return jsons;
+        return ImmutableMap.copyOf(jsons);
     }
 }